Clean client certificate auth policy issues
