+++ Packages/macOS/CmuxBrowser/Sources/CmuxBrowser/ClientCertificates/URLAuthenticationChallenge+BrowserClientCertificate.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+extension URLAuthenticationChallenge {
+    var isBrowserClientCertificateChallenge: Bool {
+        protectionSpace.authenticationMethod == NSURLAuthenticationMethodClientCertificate
+    }
+}
