+++ Packages/macOS/CmuxBrowser/Tests/CmuxBrowserTests/ClientCertificates/BrowserAuthChallengeSenderStub.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+final class BrowserAuthChallengeSenderStub: NSObject, URLAuthenticationChallengeSender {
+    func use(_ credential: URLCredential, for challenge: URLAuthenticationChallenge) {}
+    func continueWithoutCredential(for challenge: URLAuthenticationChallenge) {}
+    func cancel(_ challenge: URLAuthenticationChallenge) {}
+    func performDefaultHandling(for challenge: URLAuthenticationChallenge) {}
+    func rejectProtectionSpaceAndContinue(with challenge: URLAuthenticationChallenge) {}
+}
